+++ modules/service/rest-api-v2.js
@@ -28,116 +28,6 @@ class RestAPIServiceV2 {
         this.mapping_standards_for_event.set('graph', 'graph');
     }
 
-    /**
-     * Start RPC server
-     */
-    async startRPC() {
-        const options = {
-            name: 'RPC server',
-            version: pjson.version,
-            formatters: {
-                'application/json': (req, res, body) => {
-                    res.set('content-type', 'application/json; charset=utf-8');
-                    if (!body) {
-                        if (res.getHeader('Content-Length') === undefined && res.contentLength === undefined) {
-                            res.setHeader('Content-Length', 0);
-                        }
-                        return null;
-                    }
-
-                    if (body instanceof Error) {
-                        // snoop for RestError or HttpError, but don't rely on instanceof
-                        if ((body.restCode || body.httpCode) && body.body) {
-                            // eslint-disable-next-line
-                            body = body.body;
-                        } else {
-                            body = {
-                                message: body.message,
-                            };
-                        }
-                    }
-
-                    if (Buffer.isBuffer(body)) {
-                        body = body.toString('base64');
-                    }
-
-                    let ident = 2;
-                    if ('prettify-json' in req.headers) {
-                        if (req.headers['prettify-json'] === 'false') {
-                            ident = 0;
-                        }
-                    }
-                    const data = Utilities.stringify(body, ident);
-
-                    if (res.getHeader('Content-Length') === undefined && res.contentLength === undefined) {
-                        res.setHeader('Content-Length', Buffer.byteLength(data));
-                    }
-                    return data;
-                },
-            },
-        };
-
-        if (this.config.node_rpc_use_ssl) {
-            Object.assign(options, {
-                key: fs.readFileSync(this.config.node_rpc_ssl_key_path),
-                certificate: fs.readFileSync(this.config.node_rpc_ssl_cert_path),
-                rejectUnauthorized: true,
-            });
-        }
-
-        const server = restify.createServer(options);
-
-        server.use(restify.plugins.acceptParser(server.acceptable));
-        server.use(restify.plugins.queryParser());
-        server.use(restify.plugins.bodyParser());
-        const cors = corsMiddleware({
-            preflightMaxAge: 5, // Optional
-            origins: ['*'],
-            allowHeaders: ['API-Token', 'prettify-json', 'raw-data'],
-            exposeHeaders: ['API-Token-Expiry'],
-        });
-
-        server.pre(cors.preflight);
-        server.use(cors.actual);
-        server.use((request, response, next) => {
-            const result = this.apiUtilities.authorize(request);
-            if (result) {
-                response.status(result.status);
-                response.send({
-                    message: result.message,
-                });
-                return;
-            }
-            return next();
-        });
-
-        // TODO: Temp solution to listen all adapters in local net.
-        let serverListenAddress = this.config.node_rpc_ip;
-        if (ip.isLoopback(serverListenAddress)) {
-            serverListenAddress = '0.0.0.0';
-        }
-
-        // promisified server.listen()
-        const startServer = () => new Promise((resolve, reject) => {
-            server.listen(this.config.node_rpc_port, serverListenAddress, (err) => {
-                if (err) {
-                    reject(err);
-                } else {
-                    resolve();
-                }
-            });
-        });
-
-        await startServer(server, serverListenAddress);
-        this.logger.notify(`API exposed at  ${server.url}`);
-
-        if (this.config.is_bootstrap_node) {
-            this._exposeBootstrapAPIRoutes(server);
-        } else {
-            this._exposeAPIRoutes(server);
-        }
-    }
-
     async _check_for_handler_status(req, res) {
         const handler_object = await Models.handler_ids.findOne({
             where: {
